// web/src/api/generated/model/backendQuota.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * stocat
 * OpenAPI spec version: 0.0.1
 */
import type { BackendQuotaMode } from "./backendQuotaMode.ts";

export interface BackendQuota {
  /** @maxLength 64 */
  backend_id: string;
  /**
   * Send the limit only with the limited mode.
   * @minimum 0
   */
  limit_bytes?: number;
  mode: BackendQuotaMode;
}